requirements.txt
# Python dependencies for AI Job Matcher

# Django
Django>=4.2,<5.0
python-decouple>=3.8

# Firebase Admin SDK
firebase-admin>=6.2.0

# OpenAI
openai>=1.0.0

# CV Parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
Pillow>=10.0.0

# Production Server (optional)
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development (optional)
pytest>=7.4.0
pytest-django>=4.5.0
